+++ R/dependencies.R
@@ -592,7 +592,9 @@ check_yamls_consistent <- function(dep_structure, skip_if_missing_yaml = FALSE) 
 #' @inheritParams build_check_install
 #' @md
 #' @return A vector of 'external' R packages required to install
-#'   the selected 'internal' packages. This can be used with `remotes::system_requirements`
+#'   the selected 'internal' packages, ordered by install order (unless `from_external_dependencies`
+#'   does not include `"Depends"`, `"Imports"` and `"LinkingTo"`). The core R packages
+#'   (e.g. `methods`, `utils`) are not included. The output can be used with `remotes::system_requirements`
 #'   to extract the system requirements needed for your packages, see example below.
 #' @examples
 #' \dontrun{
@@ -682,6 +684,24 @@ get_all_external_dependencies <- function(dep_structure,
     packages_to_consider <- setdiff(new_packages_to_consider, external_packages)
   }
 
-  return(external_packages)
+  # order external dependencies by install order
+  # when ordering we use Depends, Imports and LinkingTo as they are needed for installation
+  if (!all(c("Depends", "Imports", "LinkingTo") %in% from_external_dependencies)) {
+    message("Packages will not be ordered as this requires 'Depends', 'Imports' and 'LinkingTo' to
+      included in from_external_dependencies argument.")
+  } else {
+    external_package_table <- available_packages[available_packages$Package %in% external_packages, ]
 
+    package_deps <- apply(external_package_table, 1, function(row)
+      unique(unlist(lapply(c("Depends", "Imports", "LinkingTo"), function(x) parse_deps_table(row[x]))))
+    )
+    names(package_deps) <- external_package_table$Package
+    ordered_external_packages <- topological_sort(package_deps)
+
+    # there may be packages missed in ordered_external_packages (e.g. if a package is only in suggests)
+    # so make sure they are added in
+    external_packages <- c(ordered_external_packages, setdiff(external_packages, ordered_external_packages))
+  }
+  # remove R core packages when returning the packages
+  return(external_packages[!external_packages %in% r_core_packages])
 }