+++ NEWS.md
@@ -4,6 +4,8 @@
 
 ## New feature and improvements
 
+* `get_all_external_dependencies` now returns the vector of external packages ordered by install order and the core R packages are not included. 
+
 ## Other improvements
 
 * Added a `NEWS.md` file to track changes to the package.